fix config reading

+++ src/main/java/by/nata/dao/api/IAccountDao.java
@@ -5,6 +5,6 @@ import by.nata.dto.AccountDto;
 public interface IAccountDao {
 
     AccountDto get (String accountNumber);
-    void updateAmount(int id, AccountDto accountDto);
+    void updateAmount(Long id, AccountDto accountDto);
 
 }
